added fixtures, setup and teardown
