+++ pytest.ini
@@ -1,0 +1,6 @@
+[pytest]
+testpaths=
+    "test/unit"
+    "test/selenium/"
+log_cli=true
+log_level=INFO
