+++ test/selenium/base_test.py
@@ -1,20 +1,4 @@
-from selenium import webdriver
-from selenium.webdriver.edge.options import Options
-
 class BaseTest:
-    driver = None #This is similar to a null in Java
-        
-    def setup(self):
-        print("setup")
-        if not BaseTest.driver: #singleton implementation
-            myOptions = Options()
-            BaseTest.driver = webdriver.Edge(options=myOptions)
-            self.driver = BaseTest.driver
 
-    def close_driver(self):
-        if BaseTest.driver:
-            BaseTest.driver.quit()
-            BaseTest.driver = None
-
-    def navigate(self):
-        self.driver.get("https://askomdch.com")
+    def navigate(self,driver):
+        driver.get("https://askomdch.com")
